Rx1: Move Observable Utils in another class
